+++ K8sUtils/private/Write-Header.ps1
@@ -1,16 +1,8 @@
 $script:ColorType = "ANSI"
 $script:HeaderPrefix = ""
 $script:FooterPrefix = ""
+$script:AddDate = $true
 $script:Dashes = 30
-
-function Get-TempLogFile($prefix = "k8s-") {
-    $temp = [System.IO.Path]::GetTempFileName()
-    $ret = Join-Path( Split-Path -Path $temp -Parent) "$prefix$(Split-Path -Path $temp -Leaf)"
-    $script:OutputFile = $ret
-    return $ret
-}
-
-$script:OutputFile = Get-TempLogFile
 
 function Write-MyHost {
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingWriteHost', '', Justification = 'Write-Information need ANSI resets')]
@@ -84,27 +76,27 @@ function Write-Header() {
         [int]$Length = $script:Dashes,
         [ValidateSet("None", "ANSI", "DevOps")]
         [string] $ColorType = $script:ColorType,
-        [string] $OutputFile = $script:OutputFile
+        [string] $HeaderPrefix = $script:HeaderPrefix
     )
     $headerMessage = $LogLevel -eq "error" ? "ERROR" : ""
-    $prefix = $LogLevel -eq "error" ? "" : $script:HeaderPrefix
-    Write-Status -Msg $headerMessage -LogLevel $LogLevel -ColorType $ColorType -Char '-' -OutputFile $OutputFile
-    Write-Status -Msg $msg -LogLevel normal -Length $Length -ColorType ANSI -Char ',' -OutputFile $OutputFile -Prefix $prefix -NoDate
+    $prefix = $LogLevel -eq "error" ? "" : $HeaderPrefix
+    Write-Status -Msg $headerMessage -LogLevel $LogLevel -ColorType $ColorType -Char '╒═╕' -Length 80
+    Write-Status -Msg $msg -LogLevel normal -Length $Length -ColorType ANSI -Char '─' -Prefix $prefix
+
+    $script:headerLogLevel = $LogLevel
+    $script:headerLength = $Length
+    $script:headerColorType = $ColorType
 }
 
 function Write-Footer() {
     [CmdletBinding()]
     param(
         [string]$msg,
-        [ValidateSet("error", "warning", "ok", "normal")]
-        [string] $LogLevel = "normal",
-        [int]$Length = $script:Dashes,
-        [ValidateSet("None", "ANSI", "DevOps")]
-        [string] $ColorType = $script:ColorType,
-        [string] $OutputFile = $script:OutputFile
+        [string] $FooterPrefix = $script:FooterPrefix
     )
-    $prefix = $LogLevel -eq "error" ? "" : $script:FooterPrefix
-    Write-Status -Msg $msg -LogLevel normal -Length $Length -Suffix "`n" -ColorType $ColorType -Char '`' -OutputFile $OutputFile -Prefix $prefix -NoDate
+    $prefix = $script:headerLogLevel -eq "error" ? "" : $FooterPrefix
+    Write-Status -Msg $msg -LogLevel normal -Length $script:headerLength -ColorType $script:headerColorType -Char '─' -Prefix $prefix
+    Write-Status -LogLevel $script:headerLogLevel -ColorType $script:headerColorType -Char '╘═╛' -Length 80
 }
 
 function Write-Status() {
@@ -119,19 +111,27 @@ function Write-Status() {
         [string] $Suffix = "",
         [ValidateSet("None", "ANSI", "DevOps")]
         [string] $ColorType = $script:ColorType,
-        [string] $Char = '-',
-        [string] $OutputFile = $script:OutputFile,
-        [switch] $NoDate
+        [string] $Char = '─'
     )
 
     process {
         Set-StrictMode -Version Latest
 
-        if ($NoDate) {
-            $date = ""
-        } else {
-            $date = "$((Get-Date).ToString("u")) "
+        function mapLogLevel($LogLevel) {
+            switch ($LogLevel) {
+                "error" {
+                    return "ERR"
+                }
+                "warning" {
+                    return "WRN"
+                }
+                default {
+                    return "INF"
+                }
+            }
         }
+
+        $date = $script:AddDate ? "$((Get-Date).ToString("u")) " : ""
 
         # if ($VerbosePreference -ne 'Continue') {
         $Prefix += (MapColor $LogLevel $ColorType)
@@ -142,15 +142,15 @@ function Write-Status() {
             $msgLen = ($Prefix + $date + $msg + $Suffix).Length
             if ($msgLen -lt $maxWidth) {
                 $Length = [Math]::Min($Length, $maxWidth - $msgLen - 1)
-                $msg = ($Char * $Length) + " $msg "
+                if ($Char.Length -eq 3) {
+                    $msg = ($Char[1].ToString() * ($Length-2)) + $Char[2] + " $msg "
+                } else {
+                    $msg = ($Char * $Length) + " $msg "
+                }
             }
         }
 
-        # if ($VerbosePreference -eq 'Continue') {
-        #     "${Prefix}${date}${msg}${Suffix}" | Tee-Object $OutputFile -Append | Write-Verbose
-        # } else {
-        "${Prefix}${date}${msg}${Suffix}" | Write-Plain
-        # }
+        "${Prefix}[${date}$(mapLogLevel $LogLevel)] ${msg}${Suffix}" | Write-Plain
     }
 }
 
@@ -158,11 +158,10 @@ function Write-Plain() {
     [CmdletBinding()]
     param(
         [Parameter(ValueFromPipeline)]
-        [string]$msg,
-        [string] $OutputFile = $script:OutputFile
+        [string]$msg
     )
 
     process {
-        $msg | Tee-Object $OutputFile -Append | Write-MyHost
+        $msg | Write-MyHost
     }
 }